Adding new features (sound effects, animation, score)
First of all, we have refactored the code to defragment it into pieces it helped us to organize the code.

Then we animated the homepage, using the animage module designed ealier

Next, we added a score for the user, she will now be able to know her level and focus accordingly

Lastly, we added some sound effects for fun, the user will now be able to notice by ears if she is doing well or not. In addition there are many sound effect for either success and failure